refactor: 순환 참조로 인해 solve-problem으로 exam-result를 이동
