--- feature-ui-exam-result/src/main/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Designed and developed by Duckie Team, 2022
-  ~
-  ~ Licensed under the MIT.
-  ~ Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
-  -->
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-</manifest>
