--- feature-ui-exam-result/src/main/kotlin/team/duckie/app/android/feature/ui/exam/result/ExamResultActivity.kt
@@ -1,25 +1,0 @@
-/*
- * Designed and developed by Duckie Team, 2022
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
- */
-
-package team.duckie.app.android.feature.ui.exam.result
-
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import team.duckie.app.android.feature.ui.exam.result.screen.ExamResultScreen
-import team.duckie.app.android.util.ui.BaseActivity
-import team.duckie.quackquack.ui.theme.QuackTheme
-
-class ExamResultActivity : BaseActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            QuackTheme {
-                ExamResultScreen()
-            }
-        }
-    }
-}
